[ET-35] fix(venue) : 네이밍 컨벤션에 맞게 수정 (#43)
